+++ src/Pages/Signin/index.jsx
@@ -1,0 +1,1 @@
+export {default as SigninPage} from './Signin';
