--- src/Pages/Signup/index.jsx
@@ -1,1 +1,0 @@
-export {default as SignupPage} from './Signup';
